Clean up settings menu
